First pass at a photo resizing endpoint.
This is currently fetching the images from the media_url, since I ripped
it out of another project. I still need to change it to fetch from S3
directly.

This is also using actix_web::client instead of reqwest. I should
probably switch the oauth module to do the same.